Move to_str implementations to to_str.h
